Set the user object in the request object.
