+++ server/modules/plugins/express/api.js
@@ -14,8 +14,6 @@ function plugin() {
 
     describeEndPoints: function(req, res) {
 
-      console.log(req.userObject);
-
       var plugins = _.object(_.map(plug.getPlugins(), function(plugin) {
 
         if (plugin.group != 'express') {
